fix: project structure and implemented proper paths
