+++ src/middlewares/errorHandler/errorHandler.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from 'express';
-import { ApiError } from '../../utils/ApiError.js';
+import { ApiError } from '@utils';
 import chalk from 'chalk';
 const errorHandler = (
   err: ApiError,
