+++ src/middlewares/index.ts
@@ -1,0 +1,5 @@
+import { validateRequest } from './validateRequest/validateRequest.js';
+import { upload } from './fileUpload/multer.middleware.js';
+import { errorHandler } from './errorHandler/errorHandler.js';
+import { verifyJWT } from './authHandler/auth.middleware';
+export { validateRequest, upload, errorHandler, verifyJWT };
