+++ src/models/index.ts
@@ -1,4 +1,4 @@
 import { User } from './user.model.js';
 import { Album } from './album.model.js';
-
-export { User, Album };
+import { Song } from './song.model.js';
+export { User, Album, Song };
