+++ src/modules/song/controllers/song.controller.ts
@@ -1,14 +1,13 @@
 import { Request, Response, NextFunction } from 'express';
-import { asyncHandler } from '../../../utils/asyncHandler.js';
-import SongMetaData from '../utils/songMetadata.util.js';
-import SongService from '../services/song.service.js';
 
-import { ApiError } from '../../../utils/ApiError.js';
+import { ApiError } from '@utils';
+import { SongService } from '@songModule';
 import { StatusCodes } from 'http-status-codes';
-import { Song } from '../../../models/song.model.js';
-import { ISong } from '../interfaces/song.types.interfaces.js';
-import { ApiResponse } from '../../../utils/ApiResponse.js';
-import { SongFileHash } from '../utils/songFilehash.util.js';
+import { Song } from '@models';
+import { ISong } from '@songModule';
+import { SongFileHash, SongMetaData } from '@songModule';
+import { asyncHandler } from '@utils';
+import { ApiResponse } from '@utils';
 export interface AuthenticatedRequest extends Request {
   cookies: { accessToken?: string; refreshToken?: string }; // Define cookies with accessToken
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
