+++ src/modules/user/index.ts
@@ -1,0 +1,5 @@
+import UserService from './services/userService.js';
+import JWTService from './services/jwtService.js';
+import { verifyJWT } from '@middlewares';
+import AuthController from './user.controller.js';
+export { UserService, verifyJWT, JWTService, AuthController };
