+++ src/modules/user/services/jwtService.ts
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
-import { User } from '../../../models/user.model.js';
-import { ApiError } from '../../../utils/ApiError.js';
+import { User } from '@models';
+import { ApiError } from '@utils';
 import { StatusCodes } from 'http-status-codes';
 
 class JWTService {
