+++ src/routes/album.routes.ts
@@ -1,12 +1,10 @@
 import { Router } from 'express';
-import { validateRequest } from '../middlewares/validateRequest/validateRequest.js';
-import { albumSchema } from '../modules/album/utils/albumAndSongValidation.js';
-import { uploadAlbum } from '../modules/album/middlewares/albumUpload.middleware.js';
-import { verifyJWT } from '../middlewares/authHandler/auth.middleware.js';
-import AlbumController from '../modules/album/album.controller.js';
-import { ApiError } from '../utils/ApiError.js';
+import { validateRequest } from '@middlewares';
+import { albumSchema, uploadAlbum, AlbumController } from '@albumModule';
+import { verifyJWT } from '@userModule';
+import { ApiError } from '@utils';
 import { StatusCodes } from 'http-status-codes';
-import { AuthenticatedRequest } from '../modules/song/controllers/song.controller.js';
+import { AuthenticatedRequest } from '@songModule';
 
 const router = Router();
