+++ src/routes/index.ts
@@ -1,0 +1,4 @@
+import userRouter from './user.routes.js';
+import songRouter from './song.routes.js';
+import albumRouter from './album.routes.js';
+export { userRouter, songRouter, albumRouter };
